Changes "State"->"Region", "City"->"Locality".
Some feedback from @jpmckinney:
https://github.com/bcgov/openconsult/pull/2/files/173860f49bdf2b8af20525fdb0d0338b81064de6#r59150325

>Use the more generic terms "locality" and "region" instead of "city"
and "state", which are factually incorrect for many places in Canada and
the rest of the world. Schema.org, vCard and many other standards use
those generic terms.

I'm used to trying to stick to US conventions but no need to do so here.
Updates made in the requirements and sample data.